+++ tests/run-strings/strings.cpp
@@ -161,7 +161,7 @@ void acceptStringReference(std::string & arg1)
 
 void acceptStringPointer(std::string * arg1)
 {
-    global_str = *arg1;
+  arg1->append("dog");
 }
 
 //----------------------------------------