+++ tests/strings/wrapstrings.cpp
@@ -94,7 +94,7 @@ char STR_return_char()
 }
 
 // void returnChar(char_scalar * SHF_rv +intent(out)+len(NSHF_rv))
-// function_index=21
+// function_index=22
 /**
  * \brief return a char argument (non-pointer)
  *
@@ -126,7 +126,7 @@ void STR_pass_char_ptr(char * dest, const char * src)
 }
 
 // void passCharPtr(char * dest +intent(out)+len(Ndest), const char * src +intent(in)+len_trim(Lsrc))
-// function_index=22
+// function_index=23
 /**
  * \brief strcpy like behavior
  *
@@ -163,7 +163,7 @@ void STR_pass_char_ptr_in_out(char * s)
 }
 
 // void passCharPtrInOut(char * s +intent(inout)+len(Ns)+len_trim(Ls))
-// function_index=23
+// function_index=24
 /**
  * \brief Change a string in-place
  *
@@ -196,7 +196,7 @@ const char * STR_get_char1()
 }
 
 // void getChar1(char * SHF_rv +intent(out)+len(NSHF_rv)) +pure
-// function_index=24
+// function_index=25
 /**
  * \brief return a 'const char *' as character(*)
  *
@@ -229,7 +229,7 @@ const char * STR_get_char2()
 }
 
 // void getChar2 +len(30)(char * SHF_rv +intent(out)+len(NSHF_rv))
-// function_index=25
+// function_index=26
 /**
  * \brief return 'const char *' with fixed size (len=30)
  *
@@ -262,7 +262,7 @@ const char * STR_get_char3()
 }
 
 // void getChar3(char * output +intent(out)+len(Noutput))
-// function_index=26
+// function_index=27
 /**
  * \brief return a 'const char *' as argument
  *
@@ -296,7 +296,7 @@ const char * STR_get_string1()
 }
 
 // void getString1(string & SHF_rv +intent(out)+len(NSHF_rv)) +pure
-// function_index=28
+// function_index=29
 /**
  * \brief return a 'const string&' as character(*)
  *
@@ -335,7 +335,7 @@ const char * STR_get_string2()
 }
 
 // void getString2 +len(30)(string & SHF_rv +intent(out)+len(NSHF_rv))
-// function_index=29
+// function_index=30
 /**
  * \brief return 'const string&' with fixed size (len=30)
  *
@@ -374,7 +374,7 @@ const char * STR_get_string3()
 }
 
 // void getString3(string & output +intent(out)+len(Noutput))
-// function_index=30
+// function_index=31
 /**
  * \brief return a 'const string&' as argument
  *
@@ -413,7 +413,7 @@ const char * STR_get_string2_empty()
 }
 
 // void getString2_empty +len(30)(string & SHF_rv +intent(out)+len(NSHF_rv))
-// function_index=32
+// function_index=33
 /**
  * \brief Test returning empty string reference
  *
@@ -432,7 +432,7 @@ void STR_get_string2_empty_bufferify(char * SHF_rv, int NSHF_rv)
 }
 
 // void getString5 +len(30)(string * SHF_rv +intent(out)+len(NSHF_rv))
-// function_index=33
+// function_index=34
 /**
  * \brief return a 'const string' as argument
  *
@@ -451,7 +451,7 @@ void STR_get_string5_bufferify(char * SHF_rv, int NSHF_rv)
 }
 
 // void getString6(string * output +intent(out)+len(Noutput))
-// function_index=34
+// function_index=35
 /**
  * \brief return a 'const string' as argument
  *
@@ -488,7 +488,7 @@ void STR_accept_string_const_reference(const char * arg1)
 }
 
 // void acceptStringConstReference(const std::string & arg1 +intent(in)+len_trim(Larg1))
-// function_index=36
+// function_index=37
 /**
  * \brief Accept a const string reference
  *
@@ -525,7 +525,7 @@ void STR_accept_string_reference_out(char * arg1)
 }
 
 // void acceptStringReferenceOut(std::string & arg1 +intent(out)+len(Narg1))
-// function_index=37
+// function_index=38
 /**
  * \brief Accept a string reference
  *
@@ -563,7 +563,7 @@ void STR_accept_string_reference(char * arg1)
 }
 
 // void acceptStringReference(std::string & arg1 +intent(inout)+len(Narg1)+len_trim(Larg1))
-// function_index=38
+// function_index=39
 /**
  * \brief Accept a string reference
  *
@@ -581,8 +581,40 @@ void STR_accept_string_reference_bufferify(char * arg1, int Larg1, int Narg1)
 // splicer end function.accept_string_reference_bufferify
 }
 
-// void explicit1(char * name +intent(in)+len_trim(AAlen))
+// void acceptStringPointer(std::string * arg1 +intent(inout))
 // function_index=16
+/**
+ * \brief Accept a string pointer
+ *
+ */
+void STR_accept_string_pointer(char * arg1)
+{
+// splicer begin function.accept_string_pointer
+    std::string SH_arg1(arg1);
+    acceptStringPointer(&SH_arg1);
+    strcpy(arg1, SH_arg1.c_str());
+    return;
+// splicer end function.accept_string_pointer
+}
+
+// void acceptStringPointer(std::string * arg1 +intent(inout)+len(Narg1)+len_trim(Larg1))
+// function_index=40
+/**
+ * \brief Accept a string pointer
+ *
+ */
+void STR_accept_string_pointer_bufferify(char * arg1, int Larg1, int Narg1)
+{
+// splicer begin function.accept_string_pointer_bufferify
+    std::string SH_arg1(arg1, Larg1);
+    acceptStringPointer(&SH_arg1);
+    ShroudStrCopy(arg1, Narg1, SH_arg1.c_str());
+    return;
+// splicer end function.accept_string_pointer_bufferify
+}
+
+// void explicit1(char * name +intent(in)+len_trim(AAlen))
+// function_index=17
 void STR_explicit1(char * name)
 {
 // splicer begin function.explicit1
@@ -592,7 +624,7 @@ void STR_explicit1(char * name)
 }
 
 // void explicit1(char * name +intent(in)+len_trim(AAlen))
-// function_index=39
+// function_index=41
 void STR_explicit1_BUFFER(char * name, int AAlen)
 {
 // splicer begin function.explicit1_BUFFER
@@ -606,7 +638,7 @@ void STR_explicit1_BUFFER(char * name, int AAlen)
 }
 
 // void explicit2(char * name +intent(out)+len(AAtrim))
-// function_index=17
+// function_index=18
 void STR_explicit2(char * name)
 {
 // splicer begin function.explicit2
@@ -616,7 +648,7 @@ void STR_explicit2(char * name)
 }
 
 // void explicit2(char * name +intent(out)+len(AAtrim))
-// function_index=40
+// function_index=42
 void STR_explicit2_bufferify(char * name, int AAtrim)
 {
 // splicer begin function.explicit2_bufferify
@@ -629,7 +661,7 @@ void STR_explicit2_bufferify(char * name, int AAtrim)
 }
 
 // void CreturnChar(char_scalar * SHF_rv +intent(out)+len(NSHF_rv))
-// function_index=41
+// function_index=43
 /**
  * \brief return a char argument (non-pointer), extern "C"
  *
@@ -645,7 +677,7 @@ void STR_creturn_char_bufferify(char * SHF_rv, int NSHF_rv)
 }
 
 // void CpassCharPtr(char * dest +intent(out)+len(Ndest), const char * src +intent(in)+len_trim(Lsrc))
-// function_index=42
+// function_index=44
 /**
  * \brief strcpy like behavior
  *